backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
timm
facenet-pytorch
opencv-python
numpy
Pillow
requests
beautifulsoup4
lxml
pytesseract
ollama
